Allow running mmCIF without networking
To generate the mmCIF file we query the PubMed API
(to get publication information). This won't work
when we run the test suite on our compute cluster,
since it doesn't have network access. In this case,
provide a cached copy of the publication information
so that the test can proceed.